split files and make reusable component
