+++ packages/shared/src/components/EditAltTextFlyout/index.ts
@@ -1,3 +1,4 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './EditAltTextFlyout';
+export * from './EditAltTextToolbarButton';
